Refactor of jsonSearchControllers
